requirements.txt
category_encoders==2.6.1
daal4py==2023.1.1
evidently==0.3.3
fastapi==0.100.1
fastapi_profiler==1.2.0
lightgbm==3.3.5
locustio==0.999
numpy==1.24.4
orjson==3.9.2
pandas==2.0.3
pydantic==1.10.11
pyinstrument==4.5.0
PyYAML==6.0
PyYAML==6.0.1
Requests==2.31.0
sanic==23.3.0
sanic_ext==23.3.0
scikit_learn==1.2.2
scipy==1.11.1
uvicorn==0.23.1
wandb==0.15.5
